requirements.txt
onnxruntime
opencv-python
numpy
paho-mqtt
